Refactor BoltzOutput processing and reorganize related files; add utility classes for file handling
