+++ af3_mmseqs2/processoutput/alphafold3.py
@@ -1,0 +1,7 @@
+# from af3_mmseqs2.processoutput.utils import (
+#     FileTypes,
+#     ModelCount,
+#     NpzFile,
+#     CifFile,
+#     ConfidenceJsonFile,
+# )
